naver_trend about all health_건강식품외 건강전체/requirements.txt
playwright>=1.40.0
anthropic>=0.20.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
python-dateutil>=2.8.0
reportlab>=4.0.0
youtube-transcript-api>=0.6.0
google-api-python-client>=2.100.0
adjustText>=0.8.0
